+++ pkgs/jnigen/test/kotlin_test/bindings/kotlin.dart
@@ -579,9 +579,11 @@ final class $MeasureUnit$Type$ extends jni$_.JType<MeasureUnit> {
 }
 
 /// from: `com.github.dart_lang.jnigen.Nullability$InnerClass`
-extension type Nullability$InnerClass<$T extends jni$_.JObject?,
-        $U extends jni$_.JObject, $V extends jni$_.JObject?>._(
-    jni$_.JObject _$this) implements jni$_.JObject {
+extension type Nullability$InnerClass<
+        $T extends jni$_.JObject?,
+        $U extends jni$_.JObject,
+        $V extends jni$_.JObject?>._(jni$_.JObject _$this)
+    implements jni$_.JObject {
   static final _class = jni$_.JClass.forName(
       r'com/github/dart_lang/jnigen/Nullability$InnerClass');
 
@@ -1994,7 +1996,8 @@ final class $Speed$Type$ extends jni$_.JType<Speed> {
 }
 
 /// from: `com.github.dart_lang.jnigen.SpeedUnit`
-extension type SpeedUnit._(jni$_.JObject _$this) implements jni$_.JObject {
+extension type SpeedUnit._(jni$_.JObject _$this)
+    implements jni$_.JObject, MeasureUnit {
   static final _class =
       jni$_.JClass.forName(r'com/github/dart_lang/jnigen/SpeedUnit');