+++ pkgs/jnigen/CHANGELOG.md
@@ -14,6 +14,7 @@
 - Kotlin suspend functions with no result (a return type of `Unit`) now return
   `Future<void>` in Dart instead of `Future<JObject>`.
 - Improve error message for unsupported Java class file versions in summary generation.
+- Generated extension types now implement their Java interfaces.
 
 ## 0.15.0